bpf: refactor from_netdev checks into unique ingress classifier
This commits move the current logic in from_netdev to classify a network
packet into a unique function for clarity and testing purposes.
Let's add bpf tests to these new set of functions and remove the old
`wireguard_helpers_test.c`.

While refactoring the logic into the ctx_classify_from_netdev() function,
we drop some requirement to classify WireGuard traffic. Prior to this
commit, we used to additionally verify that (a) both source and destination
L4 ports were equal to WG_PORT, and (b) that the identity was inside the
cluster. Let's now just check whether source or destination port is the
WireGuard in-use port.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>