bpf:hubble:monitor add overlay classifiers into egress path
This commits adds overlay classifier CLS_FLAG_VXLAN and CLS_FLAG_GENEVE
to the egress path of a packet, where we look at the mark to see if it
is an overlay packet. In addition, the commits adds the IsGeneve() and
IsVXLAN() function to DropNotify and TraceNotify in Monitor/Hubble.
This patch allows us to classify overlay packets in the egress path
for tracing.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>