bpf: add overlay classifiers also to ingress path
This patch makes use of the CLS_FLAG_GENEVE and CLS_FLAG_VXLAN also in the
ingress path of a packet, respectively in from_netdev.
As for marking WireGuard packets, we look at the l4 source or destination
ports, and if they match the current tunnel port.
With this, we can classify incoming overlay packets for tracing, from all
the network facing devices.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>