bpf: support additional tracing flags for trace/drop events
This commit adds the possibility to pass additional classifiers flags
to trace/drop notification events. This is useful when we compute additional
flags with respect to the base ones, such as in from_netdev where we look
at L4 ports to match WireGuard packets.
Here we also add this flags to the `struct trace_ctx` for three reasons:

1. no need to manually declare classifiers when needed, just simply
   initialize the field;
2. no need to change in future commits all the functions to accept, other
   that `struct trace_ctx` (which they already do) an additional parameter `flags`.
3. future extensions to ct lookups, to store/retrieve also classifiers
   that can be used to enhance tracing with Hubble/Monitor.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>