bpf:wireguard: land support for ingress traffic classifiers
As for bpf_host, let's use `ctx_from_netdev_classifiers{6,4}` also
in the from-wireguard hook. In previous commits we provided the support
for tracing with additional flags. Now that we have refactored the
classification logic into a unique support function, let's add it to this
WireGuard path. For the moment, the function won't set any flag, since
it only looks for WireGuard-encrypted packet, which is not observed in this
hook.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>