bpf:daemon: add TRACE_PAYLOAD_LEN_OVERLAY to node config
This commit introduces a new variable alongside the existing
TRACE_PAYLOAD_LEN. The new TRACE_PAYLOAD_LEN_OVERLAY is supposed to
be used whenever the dataplane is able to classify a packet as OVERLAY.
This is to fix cases in which Hubble/Monitor fail to decode packet due to
`TCP data offset greater than packet length` error. This error can happen
with an TCPv6 SYN (options MSS) over VXLAN: we're pushing 128B, but 144B
would be needed for correctly decoding the TCP options.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>